requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart
bcrypt==4.0.1
